app/requirements.txt
cycler==0.10.0
Django==1.11
django-cors-headers==2.1.0
django-extensions==1.9.1
django-registration-redux==1.5
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.2
matplotlib==2.0.2
mongoengine==0.13.0
nltk==3.2.4
numpy==1.7.1
pydot==1.2.3
pymongo==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.9.1
six==1.10.0
